### 2021-03-26 #3110 @dhhagan
Hi @DrDaveShaw - thank you so much for your submission to JOSS and I apologize for taking so long to get to this. Before I go ahead and find reviewers, I had a few questions/comments/suggestions:

1. It may be worth going through the README and fixing some of the typos
2. You are missing a few required parts of the paper such as the **statement of need* - it might be a good idea to look over a couple of recently published papers [here][1] to get an idea of what that looks like

Last, I was wondering if you could comment on how you envision this software to meet the "designed for maintainable extension" requirement found in the [**Substantial Scholarly Effort**][2] section of the JOSS submission criteria. It doesn't appear that any releases have been made or there is any versioning being tracked. Normally for a python submission, I would look for a python package that can be easily versioned, installed, etc, though that is **not** a strict requirement. I would like to understand how you plan on ensuring this is possible.

I look forward to evaluating the rest of this great work! Please feel free to ping me here with any comments or questions.

[1]: https://joss.theoj.org/papers/published
[2]: https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort
