### 2021-03-26 611aa4f
Update README.md
Fixing spelling mistakes and some additional clarity.
