### 2021-03-29 ece8269
Updated JOSS Paper
Updates to the JOSS paper post notes from @dhhagan, editor.
