+++ paper.bib
@@ -182,7 +182,39 @@ year = {2007}
 @article{Lakey2021,
 abstract = {The current paradigm is that indoor air constituents are uniformly mixed, with single point measurements and box modeling representing a room or house. We demonstrate that this fundamental assumption needs to be revisited through advanced model simulations and extensive measurements of bleach cleaning, which has gained particular relevance with the pandemic. We show that inorganic chlorinated products such as hypochlorous acid and chloramines exhibit spatial and vertical concentration gradients in a room and short-lived OH radicals are confined to the sunlit zone close to windows. Spatial and temporal scales of indoor constituents are modulated by rates of chemical reactions, surface interactions and building ventilation, providing critical insights for better assessments of human exposure to hazardous pollutants as well as the transport of indoor chemicals outdoors.},
 author = {Lakey, P and Youngbo, W and Shaw, D and Oesterstroem, F and Mattila, J and Reidy, E and Bottorff, B and Rosales, C. M. and Wang, C and Ampollini, L and Zhou, S and Novoselac, A and Kahan, T. F. and DeCarlo, P and Abbatt, J. P. D. and Stevens, P. S. and Farmer, D. K. and Carslaw, N. and Rim, D. and Shiraiwa, M},
-journal = {Submitted for publication in Science.},
+journal = {Submitted for publication in Science Advances.},
 title = {{Spatial and Temporal Scales of Variability for Indoor Air Constituents}},
 year = {2021}
+}
+@article{Topping2018,
+author = {Topping, David and Connolly, Paul and Reid, Jonathan},
+doi = {10.21105/joss.00755},
+issn = {2475-9066},
+journal = {J. Open Source Softw.},
+number = {28},
+pages = {755},
+title = {{PyBox: An automated box-model generator for atmospheric chemistry and aerosol simulations.}},
+volume = {3},
+year = {2018}
+}
+@article{Sommariva2020,
+author = {Sommariva, R. and Cox, S. and Martin, C. and Boro{'n}ska, K. and Young, J. and Jimack, P. K. and Pilling, M. J. and Matthaios, V. N. and Nelson, B. S. and Newland, M. J. and Panagi, M. and Bloss, W. J. and Monks, P. S. and Rickard, A. R.},
+title = {{AtChem} (version 1), an open-source box model for the {Master Chemical Mechanism}},
+journal = {Geoscientific Model Development},
+year = 2020,
+volume = 13,
+number = 1,
+pages = {169--183},
+doi = {10.5194/gmd-13-169-2020},
+}
+@article{OMeara2020,
+author = {O'Meara, Simon and Xu, Shuxuan and Topping, David and Capes, Gerard and Lowe, Douglas and Alfarra, M. and McFiggans, Gordon},
+doi = {10.21105/joss.01918},
+issn = {2475-9066},
+journal = {J. Open Source Softw.},
+number = {48},
+pages = {1918},
+title = {{PyCHAM: CHemistry with Aerosol Microphysics in Python}},
+volume = {5},
+year = {2020}
 }
