### 2021-03-29 #3110 @DrDaveShaw
Hi @dhhagan

Thank you very much for agreeing to be the Editor for our submission and for taking the time to look it over and provide these questions/comments/suggestions.

1. I have read over the README and I have fixed the typos. I'm sorry they made it through to submission!
2. I have extracted the _statement of need_ into a section and completed additions to the paper in line with the recently published papers linked in your reply.

Prior to these edits I have released a version (v1.0) of the software. Not releasing a version prior to review was an oversight that I can only ascribe to my lack of experience with the process. As the software is further developed we will continue to release versions following semantic versioning guidelines and with major releases being accompanied by appropriate publications. Hopefully this is sufficient to meet the requirement for "designed for maintainable extension".

Thank you again for taking the time for these comments and I look forward to the review process.
