### 2021-05-01 #3224 @goldmanm
@DrDaveShaw. I really enjoyed testing this new package for indoor air chemistry that you wrote. I appreciate you taking the time to share it with the wider community.

A few initial questions/comments:

readme

1. Typically people often want improvements to the code to be submitted as a pull request. In the readme file, you mention having people email updates, which might not be the most efficient method of improving the code. I would recommend requesting changes be submitted as a pull request, if applicable. And if not applicable, to have the modified code be hosted on a public repository in a timely fashion.

user guide

1. Would it be possible to include source, .tex, .docx, etc, for the user manual in the repository itself so that others can adjust and build upon it if they notice any issues? You could create a separate folder in the repository for the source code.
2. The repository uses GNUv3 while the user manual uses GNUv1.3. Is it possible to synchronize these to the same license?
3. It might be useful to specify a set of versions of the python packages which work with the code, in case there are future package incompatibility issues.
4. I am not sure of the statement: "If you use Anaconda, then the install includes all of the packages required and no further
downloads or installs are required." I had to install and download a few of these packages with `conda install numba` even though I have anaconda installed on my computer.
5. Some description of how to implement additional chemistries into the code should be written in section 7.6, since I anticipate users might want to modify chemistry.

manuscript

1. While the manuscript mentions what the software can do, it might also be useful to be more explicit about its current limitations. Is it able to resolve concentrations spatially within a room? Can it model multiple connected indoor spaces? Does any chemistry happen in or on the particle phase? Are there limitations to how wall-effects are accounted for? I can infer some of these from the user manual, but it isn't explicitly mentioned.
2. Remove 'easily explored' from the end of the statement of need section.

model setup and use

1. Just for my edification, when the user sets `M` in `custom_input.txt`, the number density of air includes or excludes the additional VOCs being studied?
2. The default settings.py file which comes with the software tkes about 90 minutes to run. Was there any reason that this particular set of conditions was chosen to be the default that is deployed, if it is (for example, the run gives an output used in a publication), that might be useful to mention in the user guide. If it is not, do you forsee any issue with making the run shorter, so people can more quickly see if the code is functioning and analyze the output?
4. I modified the `INCHEM_extractor.py` file to use the updated path of the folder path from my run, and when I ran `INCHEM_extractor.py`, I got an error message.
    ```
    KeyError: "Passing list-likes to .loc or [] with any missing labels is no longer supported. The following labels were missing: Index(['HO2_reactivity', 'HO2_production'], dtype='object').
    ```
    Changing HO2_production and HO2_reactivity to OH_production and OH_reactivity allowed for INCHEM_extractor.py to execute successfully. Was there something I did wrong when running or analyzing the code? Why was the column of data for 'HO2_reactivity' not in 'out_data.pickle'?
4. Let's say I am a user who downloads and runs the software according to the manual; how can I be confident that my code is working as expected (and some different version of numpy or numba didn't break the code)? You might want to add statements to the user guide that make it explicit what attributes the user should check to ensure the code is properly functioning.
