### 2021-05-04 #3224 @DrDaveShaw
@goldmanm. Thank you very much for taking the time to begin a review of our package, your comments and questions are exceptionally helpful. I'll be going through them individually over the next couple of days making suggested adjustments or additions and will have a full response for you shortly.
