### 2021-05-04 738f86d
Update README.md
Improvements to the code now to be submitted as a pull request. Response to JOSS review.
