### 2021-05-05 5b3eb9c
Update custom_input.txt
Rate to rate coefficient for accuracy
