+++ custom_input.txt
@@ -23,10 +23,10 @@
 #You should have received a copy of the GNU General Public License
 #along with INCHEM-Py.  If not, see <https://www.gnu.org/licenses/>.
 #*****************************************************************************
-# Custom reaction input file. All lines that are not rates or reactions
-# must start with a hash (#) and thus be commented out. 
+# Custom reaction input file. All lines that are not rate coefficients or 
+# reactions must start with a hash (#) and thus be commented out. 
 #
-# Rates must be put in in the form:
+# Rates coefficients must be put in in the form:
 # name = calculation
 #
 # Reactions must be in the form:
