### 2021-05-07 #3224 @khinsen
@dhhagan After a first look at the paper and the software, I wonder what exactly JOSS expects me to do to "confirm the functional claims of the software".

The claim is basically that the program "creates and solves a system of coupled Ordinary Differential Equations (ODEs) to progress indoor atmospheric chemical species concentrations through time." Looking at the code suggests that it does indeed set up and solve a system of ODEs. But I have no idea what those equations should be, because all the scientific papers cited are behind a paywall to which I have no access. For the same reason, I cannot judge either if the numerical methods employed seem appropriate to these specific equations.
