### 2021-05-10 22f2a55
Added warning
From JOSS review, added a warning about adding parameters that already have names.
