### 2021-05-10 cd2c30d
Added manual source files
Response to JOSS review, manual source files added.
