+++ manual source files/INCHEM_manual.tex
@@ -1,0 +1,998 @@
+% %
+% LAYOUT_E.TEX - Short description of REFMAN.CLS
+%                                       99-03-20
+%
+%  Updated for REFMAN.CLS (LaTeX2e)
+%
+\documentclass[a4paper]{refart}
+\usepackage{makeidx}
+\usepackage{ifthen}
+\usepackage{graphicx}
+\usepackage{amsmath}
+\usepackage{amsfonts}
+\usepackage{tabularx}
+\usepackage{url}
+\usepackage{hyperref}
+\hypersetup{
+    colorlinks=true,
+    linkcolor=blue,
+    filecolor=magenta,      
+    urlcolor=blue,
+    citecolor=blue
+}
+\usepackage{dirtree}
+\usepackage[labelfont=bf]{caption}
+
+\usepackage[backend=biber,style=numeric,
+            sorting=none, natbib=true]{biblatex} % Complete re-implementation of bibliographic facilities
+\addbibresource{references.bib}
+
+\def\bs{\char'134 } % backslash in \tt font.
+\newcommand{\ie}{i.\,e.,}
+\newcommand{\eg}{e.\,g..}
+\DeclareRobustCommand\cs[1]{\texttt{\char`\\#1}}
+
+\title{Indoor Chemical Model (INCHEM-Py v1.x) \\
+User Manual\\
+\begin{figure}
+    \hspace{-0.25\textwidth}\hspace{0.25in}
+    \includegraphics[width=\textwidth]{INCHEMPY_logo.png}
+\end{figure}}
+
+\author{
+David Shaw\\
+\texttt{david.shaw@york.ac.uk}\\
+Nic Carslaw\\
+\texttt{nicola.carslaw@york.ac.uk}\\
+}
+
+\date{Updated: \today}
+\emergencystretch1em  %
+
+\pagestyle{myfootings}
+\markboth{INCHEM-Py}%
+         {INCHEM-Py}
+
+\makeindex 
+
+\setcounter{tocdepth}{2}
+
+\begin{document}
+
+\maketitle
+
+
+%\begin{abstract}
+%abstract
+%\end{abstract}
+\newpage
+\tableofcontents
+
+
+\vspace*{\fill}
+
+Copyright \textcopyright \hspace{1mm}2019-2021 \\
+David Shaw : david.shaw@york.ac.uk\\
+Nicola Carslaw : nicola.carslaw@york.ac.uk\\
+
+
+Permission is granted to copy, distribute and/or modify this
+document under the terms of the GNU Free Documentation License,
+Version 1.3 or any later version published by the Free Software
+Foundation; with no Invariant Sections, no Front-Cover Texts, and
+no Back-Cover Texts.  A copy of the license is available at \url{https://www.gnu.org/licenses/fdl-1.3.html}.
+
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+\newpage
+\section{INCHEM-Py overview}
+The INdoor CHEMical model in Python (INCHEM-Py) is an open source box model that creates and solves a system of coupled Ordinary Differential Equations (ODEs) to provide predicted concentrations of indoor air pollutants through time. It is a refactor of the indoor detailed chemical model, developed by Carslaw  \cite{Carslaw2007}, with improvements in form, function, and accessibility.
+
[…]
