### 2021-05-10 3fb567a
Update INCHEM_main.py
Added an additional check to compare custom inputs with predefined variables. Will now ignore any custom input that clashes with a pre-defined constant and will produce a print statement to make the user aware an input has been ignored.
