### 2021-05-10 0b34911
Update settings.py
Default values changes to simulate a single day reducing default model run time.
