### 2021-05-10 #3224 @dhhagan
@DrDaveShaw Are you able to provide copies of the relevant cited papers to @khinsen for the purpose of reviewing this paper?
