### 2021-05-11 #3224 @DrDaveShaw
@goldmanm. Thank you again for your comments and questions. I have been through them and my responses are below with references to any changes I have made to the package.

readme
1. >Typically people often want improvements to the code to be submitted as a pull request. In the readme file, you mention having people email updates, which might not be the most efficient method of improving the code. I would recommend requesting changes be submitted as a pull request, if applicable. And if not applicable, to have the modified code be hosted on a public repository in a timely fashion.

    I have changed the readme and the manual to *"...we ask that you submit any improvements you make to the code to us by submitting a pull request."* We do want the code to be improved by the community and this process improvement will work towards that.

user guide
1. >Would it be possible to include source, .tex, .docx, etc, for the user manual in the repository itself so that others can adjust and build upon it if they notice any issues? You could create a separate folder in the repository for the source code.

    A folder containing the source code for the user manual has been added to the repository.

2. >The repository uses GNUv3 while the user manual uses GNUv1.3. Is it possible to synchronize these to the same license?

    These refer to different licenses. The GNU General Public License for software is currently at version 3 but the manual is under the GNU Free Documentation License which is currently at version 1.3. This is as per the recommendations given on the [GNU website.](https://www.gnu.org/licenses/licenses.html)

3. >It might be useful to specify a set of versions of the python packages which work with the code, in case there are future package incompatibility issues.

    I have added the known working versions of the python packages to the dependencies section of the manual.

4. >I am not sure of the statement: "If you use Anaconda, then the install includes all of the packages required and no further
downloads or installs are required." I had to install and download a few of these packages with 'conda install numba' even though I have anaconda installed on my computer.

    I agree that this is a potentially inaccurate and rather broad statement. Currently all packages required are included in the [Anaconda 2020.11](https://docs.anaconda.com/anaconda/reference/release-notes/) version and the install has been tested. Given that this has historically not been the case and that previous versions of Anaconda will not have included all of the packages I have made the following adjustment to the text in the manual:

    *INCHEM-Py relies on a number of Python packages. If you use Anaconda, then the current version at time of writing (Anaconda 2020.11) includes all of the packages required and no further downloads or installs are required.*

    *If you have opted to not use Anaconda, are using an older version, or you have opened a new environment without the default packages, then the following packages need to be installed:*

    Hopefully this is significantly clearer and now accurate.

5. >Some description of how to implement additional chemistries into the code should be written in section 7.6, since I anticipate users might want to modify chemistry.

	In section 7.6 we describe the additional indoor chemistry that has been developed and used by Carslaw and her group and which is not included in the Master Chemical Mechanism (MCM). This chemistry can be modified by users as long as the format of the module is maintained and as such I have added a description of the formatting of this module to section 7.6. However users should not add additional chemistry here, the custom\_input.txt file has been included for this purpose and its use is described in section 4.3. A link to this section with an explanation has also now been added to section 7.6 for clarity.

	I have also made a slight adjustment to the text in section 7.5 to describe adding surface deposition for a new species in the custom\_input.txt file with an example.

manuscript
1. >While the manuscript mentions what the software can do, it might also be useful to be more explicit about its current limitations. Is it able to resolve concentrations spatially within a room? Can it model multiple connected indoor spaces? Does any chemistry happen in or on the particle phase? Are there limitations to how wall-effects are accounted for? I can infer some of these from the user manual, but it isn't explicitly mentioned.

    I have added additional text to the manuscript stating: *[INCHEM-Py] "does not solve for spatial dimensions and assumes a well mixed atmosphere."* I have also added further text and citations that link the methods methods for representing both the gas-to-particle partitioning for three monoterpenes and the surface deposition for relevant species, which are both discussed in detail in Carslaw 2012. Some of these methods require further reading as full explanations are beyond the scope of this paper, but we have provided enough detail for a user to decide if INCHEM-Py is suitable for their needs or can be further developed to be so.

2. >Remove 'easily explored' from the end of the statement of need section.

	Done.

model setup and use
1. >Just for my edification, when the user sets M in custom\_input.txt, the number density of air includes or excludes the additional VOCs being studied?

   I assume you mean setting M in the settings.py file? M is the number density of all molecules in 1 cm<sup>3</sup> of air and should be calculated for the temperature and pressure required.

   Given that this question revealed that M could be set in the custom\_input.txt file I have added a warning to try and present issues of users trying to set constants in this way.

    "Be careful not to add any reactions that have already been included in the model as adding them twice will duplicate them and therefore double their rate. Any new species must have a unique name. Do not attempt to redefine any constants within this file, they can be defined in the settings.py file."

    I have also added a check within INCHEM_main.py that will check the rate coefficients (the only way to define a single value in the custom\_inputs.txt file) for any constants that have already been defined and ignore them with the following print statement which should prevent this issue in the future.

    "x from custom_input.txt ignored. Defined elsewhere."

2. >The default settings.py file which comes with the software tkes about 90 minutes to run. Was there any reason that this particular set of conditions was chosen to be the default that is deployed, if it is (for example, the run gives an output used in a publication), that might be useful to mention in the user guide. If it is not, do you forsee any issue with making the run shorter, so people can more quickly see if the code is functioning and analyze the output?

     We agree that reducing the length of time simulated would decrease the run time and allowing users to be able to compare and check function in a shorter amount of time would be beneficial. We typically run the model for 3 days to ensure that steady-state is attained.

    I have reduced the default value of "seconds\_to\_integrate" to 86400 which is one day and will take ~30 minutes to run. I have also updated the relevant sections of the manual.

3. >I modified the INCHEM\_extractor.py file to use the updated path of the folder path from my run, and when I ran INCHEM\_extractor.py, I got an error message. Changing HO2\_production and HO2\_reactivity to OH\_production and OH\_reactivity allowed for INCHEM\_extractor.py to execute successfully. Was there something I did wrong when running or analyzing the code? Why was the column of data for 'HO2\_reactivity' not in 'out_data.pickle'?

    This is a mistake on our part as we removed all but OH production and reactivity calculations from the default model. Instructions on how to add more species are included in the manual but HO<sub>2</sub> reactivity and production are not calculated in the model by default. These have now been changed these to OH in the INCHEM\_extractor.py script.

4. >Let's say I am a user who downloads and runs the software according to the manual; how can I be confident that my code is working as expected (and some different version of numpy or numba didn't break the code)? You might want to add statements to the user guide that make it explicit what attributes the user should check to ensure the code is properly functioning.

    To check model function, we have included the values for simulated  O<sub>3</sub> concentrations from the default version of the model in a csv format in the test files folder. Any change in the function of the packages will change the outputs of the model.We chose ozone as it has many links to chemical and physical transformations in the model and is a good indicator species. Clearly though, the user would need to run the model in the default condition to be able to reproduce the ozone concentrations. A test script already exists to test the validity of the functions that build the system of ODEs. I have added this method of testing model function into the manual.

Thank you again for taking the time to review our software and suggest useful improvements.
