### 2021-05-11 570c82e
Updated paper (JOSS)
Updates to the paper detailing some limitations and linking methods. Edit in response to initial JOSS review.
