+++ paper.bib
@@ -68,7 +68,7 @@ volume = {240},
 year = {2020}
 }
 @article{Wong2017,
-abstract = {Ambient levels of chlorinated gases and aerosol components were measured by online chemical ionization and aerosol mass spectrometers after an indoor floor were repeatedly washed with a commercial bleach solution. Gaseous chlorine (Cl2, 10's of ppbv) and hypochlorous acid (HOCl, 100's of ppbv) arise after floor washing, along with nitryl chloride (ClNO2), dichlorine monoxide (Cl2O), and chloramines (NHCl2, NCl3). Much higher mixing ratios would prevail in a room with lower and more commonly encountered air exchange rates than that observed in the study (12.7 h−1). Coincident with the formation of gas-phase species, particulate chlorine levels also rise. Cl2, ClNO2, NHCl2, and NCl3 exist in the headspace of the bleach solution, whereas HOCl was only observed after floor washing. HOCl decays away 1.4 times faster than the air exchange rate, indicative of uptake onto room surfaces, and consistent with the well-known chlorinating ability of HOCl. Photochemical box modeling captures the temporal profiles of Cl2 and HOCl very well and indicates that the OH, Cl, and ClO gas-phase radical concentrations in the indoor environment could be greatly enhanced ({\textgreater}106 and 105 cm−3 for OH and Cl, respectively) in such washing conditions, dependent on the amount of indoor illumination.},
+abstract = {Ambient levels of chlorinated gases and aerosol components were measured by online chemical ionization and aerosol mass spectrometers after an indoor floor were repeatedly washed with a commercial bleach solution. Gaseous chlorine (Cl2, 10's of ppbv) and hypochlorous acid (HOCl, 100's of ppbv) arise after floor washing, along with nitryl chloride (ClNO2), dichlorine monoxide (Cl2O), and chloramines (NHCl2, NCl3). Much higher mixing ratios would prevail in a room with lower and more commonly encountered air exchange rates than that observed in the study (12.7 h−1). Coincident with the formation of gas-phase species, particulate chlorine levels also rise. Cl2, ClNO2, NHCl2, and NCl3 exist in the headspace of the bleach solution, whereas HOCl was only observed after floor washing. HOCl decays away 1.4 times faster than the air exchange rate, indicative of uptake onto room surfaces, and consistent with the well-known chlorinating ability of HOCl. Photochemical box modeling captures the temporal profiles of Cl2 and HOCl very well and indicates that the OH, Cl, and ClO gas-phase radical concentrations in the indoor environment could be greatly enhanced ({\textgreater}106 and 105 cm−3 for OH and Cl, respectively) in such washing conditions, dependent on the amount of indoor illumination.},
 author = {Wong, J. P.S. and Carslaw, N. and Zhao, R. and Zhou, S. and Abbatt, J. P.D.},
 doi = {10.1111/ina.12402},
 file = {:A$\backslash$:/Downloads (A)/ina.12402.pdf:pdf},
@@ -125,7 +125,7 @@ volume = {80},
 year = {2013}
 }
 @article{Kruza2017,
[…]
