+++ paper.md
@@ -26,11 +26,13 @@ In developed countries people spend over 90% of their time indoors where they ar
 
 # Statement of need
 
-Over the last 15 years, the INdoor Detailed Chemical Model (INDCM) has been used to successfully probe the chemistry of indoor air [@Carslaw2007]. However, it relies on proprietary software and requires a high level of chemistry expertise and some Fortran knowledge to edit and use the code. Software tools such as Pybox [@Topping2018], PyCHAM [@OMeara2020] and AtChem2 [@Sommariva2020] facilitate the use of chemical mechanisms to model atmospheric chemistry, but with a focus on chamber studies or ambient conditions. INCHEM-Py has been designed with a unique set of tools for the specific purpose of modelling indoor air chemistry. As well as a detailed gas-phase chemical mechanism, the new model includes gas-to-particle partitioning for three of the commonly encountered terpenes indoors (limonene and alpha- and beta-pinene), novel indoor photolysis parameterisation, indoor-outdoor air exchange and deposition to internal surfaces. INCHEM-Py is open source, has no black box processes and all inputs can be tracked through the model, allowing for complete understanding of the system. It has been designed to be easy to install for use by academics and students of all abilities, and is sufficiently accessible for further development by the wider indoor air community. The functionality embedded within INCHEM-Py will allow for a wide range of uses including in-depth analysis of experimental measurements, development and testing of new chemical mechanisms and probing numerous indoor scenarios, with the impacts on simulated indoor air pollutant concentrations from variations in parameters such as photolysis, ventilation and deposition rates, outdoor pollutant concentratons, time of year, and building location easily explored. 
+Over the last 15 years, the INdoor Detailed Chemical Model (INDCM) has been used to successfully probe the chemistry of indoor air [@Carslaw2007]. However, it relies on proprietary software and requires a high level of chemistry expertise and some Fortran knowledge to edit and use the code. Software tools such as Pybox [@Topping2018], PyCHAM [@OMeara2020] and AtChem2 [@Sommariva2020] facilitate the use of chemical mechanisms to model atmospheric chemistry, but with a focus on chamber studies or ambient conditions. INCHEM-Py has been designed with a unique set of tools for the specific purpose of modelling indoor air chemistry. As well as a detailed gas-phase chemical mechanism, the new model includes gas-to-particle partitioning for three of the commonly encountered terpenes indoors (limonene and alpha- and beta-pinene), novel indoor photolysis parameterisation, indoor-outdoor air exchange and deposition to internal surfaces. INCHEM-Py is open source, has no black box processes and all inputs can be tracked through the model, allowing for complete understanding of the system. It has been designed to be easy to install for use by academics and students of all abilities, and is sufficiently accessible for further development by the wider indoor air community. The functionality embedded within INCHEM-Py will allow for a wide range of uses including in-depth analysis of experimental measurements, development and testing of new chemical mechanisms and probing numerous indoor scenarios, with the impacts on simulated indoor air pollutant concentrations from variations in parameters such as photolysis, ventilation and deposition rates, outdoor pollutant concentrations, time of year, and building location. 
 
 # INCHEM-Py
 
-INCHEM-Py creates and solves a system of coupled Ordinary Differential Equations (ODEs) to progress indoor atmospheric chemical species concentrations through time. It can be used to investigate numerous indoor air chemistry events in world-wide locales providing key insights into the chemical processes involved. INCHEM-Py utilises the Master Chemical Mechanism (MCM) at its core [@Jenkin1997], which is near explicit and contains no lumping or use of species surrogates.
[…]
