### 2021-05-13 4cf60de
Update particle_input.py
Modified scaling factor based on Carslaw 2012
