### 2021-05-13 2968113
Update particle_input.py
Fixed an unnecessary duplication.
