### 2021-05-14 fdc0bf7
Update particle_input.py
Value update. SCALINGFAC is set to 1 as it is not required but may be required for specific applications of the model in future.
