### 2021-05-18 d4d56b7
Closes #6
The use of `global` has been removed from the functions within `modules.INCHEM_main`.
