### 2021-05-18 83b5e51
run_inchem function created. Closes #3
Removes circular import
