+++ modules/INCHEM_main.py
@@ -24,916 +24,911 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with INCHEM-Py.  If not, see <https://www.gnu.org/licenses/>.
 """
-
-"""
-import all user set variables
-"""
-import sys
-sys.path.append("..") #needed to import the variables from settings.py
-
-from settings import filename, particles, INCHEM_additional, custom, temp,\
-    rel_humidity, M, const_dict, AER, diurnal, city, date, lat, light_type,\
-        light_on_times, glass, HMIX, initials_from_run, initial_conditions_gas,\
-            timed_emissions, timed_inputs, dt, t0, seconds_to_integrate,\
-                custom_name, output_graph, output_species
+def run_inchem(filename, particles, INCHEM_additional, custom, temp, rel_humidity,
+               M, const_dict, AER, diurnal, city, date, lat, light_type, 
+               light_on_times, glass, HMIX, initials_from_run,
+               initial_conditions_gas, timed_emissions, timed_inputs, dt, t0,
+               seconds_to_integrate, custom_name, output_graph, output_species):
   
-'''
-import all modules
-'''
-import pickle
-from modules.Import import import_all, custom_import 
-from modules.particle_input import particle_import, particle_calcs, reactions_check
-from modules.photolysis import photolysis_J, Zixu_photolysis, Zixu_photolysis_compiled
-from modules.initial_dictionaries import initial_conditions, master_calc, master_compiler,\
-    reaction_rate_compile, reaction_eval, write_jacobian_build, INCHEM_species_calc
-from modules.outdoor_concentrations import outdoor_rates, outdoor_rates_diurnal, outdoor_rates_calc
-import numpy as np
-import numba as nb
-from scipy.integrate import ode
-import time
-from modules.surface_dictionary import surface_deposition
-from threadpoolctl import threadpool_limits
-import importlib.util
-import pandas as pd
-import os
-import datetime
-import math
-import time as timing
-from modules.reactivity import reactivity_summation, reactivity_calc, production_calc
-
-sys.setrecursionlimit(4000) #to compile the master array the recursion limit
-#must be increased as some of the evaluated ODEs are greater than the usual
-#system limit
-
-save_rate = 1 #sets the rate at which outputs are saved within the integrator.
-#Useful if the timestep has to be reduced but an output at a specific interval
-#is still required. A save rate of 1 will save every dt, a save rate of 2 will
-#save every 2*dt    
-
-def summations_compile(sums):
     '''
-    compiles any custom summations
+    import all modules
+    '''
+    import sys
+    import pickle
+    from modules.Import import import_all, custom_import 
+    from modules.particle_input import particle_import, particle_calcs, reactions_check
+    from modules.photolysis import photolysis_J, Zixu_photolysis, Zixu_photolysis_compiled
+    from modules.initial_dictionaries import initial_conditions, master_calc, master_compiler,\
+        reaction_rate_compile, reaction_eval, write_jacobian_build, INCHEM_species_calc
+    from modules.outdoor_concentrations import outdoor_rates, outdoor_rates_diurnal, outdoor_rates_calc
+    import numpy as np
+    import numba as nb
+    from scipy.integrate import ode
+    import time
+    from modules.surface_dictionary import surface_deposition
+    from threadpoolctl import threadpool_limits
+    import importlib.util
+    import pandas as pd
+    import os
+    import datetime
+    import math
+    import time as timing
+    from modules.reactivity import reactivity_summation, reactivity_calc, production_calc
     
-    inputs:
-        sums = list of summations [name of sum, calculation]
+    sys.setrecursionlimit(4000) #to compile the master array the recursion limit
+    #must be increased as some of the evaluated ODEs are greater than the usual
[…]
