### 2021-05-18 7d278dd
Particle import bug fixed
Imported values of seed and seed_1 were being overwritten. Now they are only set to initial values if no imported values exist.
