### 2021-05-19 #3224 @goldmanm
@DrDaveShaw Thank you for implementing all these changes. I am going to double check documentation creation and rerun the example and analysis to make sure it is all functioning properly.

Towards the end of this process, it might be useful to create a new version number and mention this version number in the manuscript for the package (so users can easily know exactly what software version the manuscript corresponds to, in case additional chemistry or functionality is added later). Maybe v1.1 or v1.0.1 or v2, your choice.

I'm also considering running some other cases to double check that what I expect to occur is what occurs.
