### 2021-05-19 #3224 @DrDaveShaw
@goldmanm I believe we'll go to v1.1 after this review process so I will change the manuscript accordingly.

I am still working my way through issues raised by @khinsen on the repository itself so I don't know if you would prefer to wait until they are complete to run your checks?

Thank you again and let me know if there's anything else you would like us to address.
