### 2021-05-19 c31e3ce
Update paper.md
Added a statement about the version the publication is referring to/will be referring to when the JOSS review is complete. (v1.1)
