+++ paper.md
@@ -34,7 +34,9 @@ INCHEM-Py creates and solves a system of coupled Ordinary Differential Equations
 
 ICHEM-Py does not solve for spatial dimensions and assumes a well mixed atmosphere [@Carslaw2007]. Gas-to-particle partitioning is implemented using absorptive partitioning from @Pankow1994 with relevant reactions between particle and gas phase species included within the mechanism [@Carslaw2012]. Surface deposition is treated as an irreversible process with rates dependant on the simulated surface area to volume ratio and species deposition velocities [@Carslaw2012]. Photolysis rates can be calculated for several indoor lighting sources with different spectral profiles and also for attenuation of sunlight through multiple glass compositions [@Wang2021]. Daylight hours are derived from latitude and time of year.
 
-The additional chemical mechanisms developed specifically for indoor air scenarios contained within the model have been utilised in a numerous published studies. These include: indoor air chemistry following cleaning with terpene based mixtures [@Carslaw2012; @Carslaw2017; @Carslaw2013; @Terry2014]; indoor air chemistry following cleaning with chlorine containing bleach [@Wong2017]; the impact of outdoor vegetation on indoor air chemistry [@Carslaw2015]; the importance of surface interactions for secondary pollutant formation [@Kruza2017]; ranking of harmful volatile organic compounds (VOCs) [@Carslaw2019]; improved model representation of the formation and composition of aerosols [@Kruza2020]; and photolysis of indoor air chemistry following high-concentration hospital/industrial cleaning events [@Wang2020]. INCHEM-Py has already been used to determine production rates and reactivity of indoor radical species, to assess the spatial and temporal scales of variability for indoor air constituents [@Lakey2021], and is currently being used to probe the impact of indoor air chemistry on ambient air, as well as to compare the differential secondary pollutant formation potential for different cleaning formulations.  
+The additional chemical mechanisms developed specifically for indoor air scenarios contained within the model have been utilised in a numerous published studies. These include: indoor air chemistry following cleaning with terpene based mixtures [@Carslaw2012; @Carslaw2017; @Carslaw2013; @Terry2014]; indoor air chemistry following cleaning with chlorine containing bleach [@Wong2017]; the impact of outdoor vegetation on indoor air chemistry [@Carslaw2015]; the importance of surface interactions for secondary pollutant formation [@Kruza2017]; ranking of harmful volatile organic compounds (VOCs) [@Carslaw2019]; improved model representation of the formation and composition of aerosols [@Kruza2020]; and photolysis of indoor air chemistry following high-concentration hospital/industrial cleaning events [@Wang2020]. INCHEM-Py has already been used to determine production rates and reactivity of indoor radical species, to assess the spatial and temporal scales of variability for indoor air constituents [@Lakey2021], and is currently being used to probe the impact of indoor air chemistry on ambient air, as well as to compare the differential secondary pollutant formation potential for different cleaning formulations.
+
+At publication the current stable release of INCHEM-Py is v1.1.  
 
 # Acknowledgements
 
