### 2021-05-19 f9e0827
All binary choices adjusted to boolean
All binary choices from settings.py and elsewhere have been adjusted to boolean as recommended in #4

Documentation needs to be updated so issue remains open.
