### 2021-05-19 #3224 @khinsen
@DrDaveShaw @dhhagan Sorry for the delay - here comes my review! Thanks to the thorough first review by @goldmanm, the current version of code and paper are already much improved, which simplified my task.

INCHEM-Py belongs to the rather peculiar category of scientific software called "computational models". It is not a program/library that *implements* a model, it *is* the model. In theory, the model is a large set of differential equations with lots of parameters, but it's too large to be written down in any other useful form than software that computes solutions of the equations. The JOSS reviewer checklist needs some interpretation for computational models. The "functionality" it refers to is rather simple: run a simulation. There isn't much of an API to document or test. The real value of this software is the science it encodes. But the science is hard to document precisely, and even harder to test.

The provided documentation does a good job of describing what the software does, but readers will have to consult at least the reference Carslaw (2007) to understand how the model works. Unfortunately this is a paywalled article. Potential users who want to convince themselves that the software actually implements the science correctly will have to read the source code (as for any computational model). For this reason, I went through the code and opened a few issues on the repository to suggest improvements to readability. Note however that I did not check conformity to the description in the paper cited above, as that would be a huge effort. Validity checks will have to be made continuously by the user community, and for that reason I can only applaud the authors' decision to thoroughly document and publish the code.

Concerning the software paper, I have nothing to add to @goldmanm's suggestions, which are already implemented in the current version.
