### 2021-05-19 #3224 @DrDaveShaw
@khinsen thank you very much for your review, we appreciate the time taken and the level of detail given with the issues raised. I am working through these readability issues now and will have them complete for version 1.1 of the model to accompany the  paper.

With regards to the workings of the model and consulting Carslaw (2007) and other referenced/paywalled works. It is our intention to write up fully the science modelled in detail, in a single open access publication, as we are aware that its development has been over a number of years and a number of papers. This will provide a greater level of accessibility and confidence in the software, hopefully growing the community of users.
