### 2021-05-19 fcf9c6e
Updated particle implementation
Removing `global` during issue #6 resulted in the code not updating the tsp and tspx calculations. Instead I have removed the particle dictionary from being passed around and instead it now updates the density dictionary. This is more efficient and more accessible.
