+++ modules/INCHEM_main.py
@@ -138,7 +138,7 @@ def run_inchem(filename, particles, INCHEM_additional, custom, temp, rel_humidit
     
     
     def dy_calc(master_compiled,reaction_rate_dict,calc_dict,density_dict,outdoor_dict,\
-                surface_dict,particle_dict,species,timed_dict):
+                surface_dict,species,timed_dict):
         '''
         evaluating the master array of species
         
@@ -149,7 +149,6 @@ def run_inchem(filename, particles, INCHEM_additional, custom, temp, rel_humidit
             density_dict = dictionary of current concentrations
             outdoor_dict = dictionary of outdoor species concentrations
             surface_dict = dictionary of surface deposition rates for each species
-            particle_dict = dictionary of particle constants and variables for particle calculations
             species = list of species
             timed_dict = dictionary of rates for any timed inputs for current time
             
@@ -158,7 +157,7 @@ def run_inchem(filename, particles, INCHEM_additional, custom, temp, rel_humidit
         '''
         dy_dict={}    
         full_dict={**reaction_rate_dict,**calc_dict,**density_dict,**outdoor_dict,\
-                   **surface_dict,**particle_dict,**timed_dict}
+                   **surface_dict,**timed_dict}
         
         for specie in species:
             dy_dict[specie]=(eval(master_compiled[specie],{},full_dict))
@@ -166,7 +165,7 @@ def run_inchem(filename, particles, INCHEM_additional, custom, temp, rel_humidit
     
     
     def dy_dy_calc(dy_dy_dict,J_dict,calc_dict,density_dict,species,outdoor_dict,\
-                   surface_dict,particle_dict,num_species,timed_dict,reaction_rate_dict):
+                   surface_dict,num_species,timed_dict,reaction_rate_dict):
         '''
         Evaluating the Jacobian during integration
         
@@ -178,7 +177,6 @@ def run_inchem(filename, particles, INCHEM_additional, custom, temp, rel_humidit
             density_dict = dictionary of current concentrations
             outdoor_dict = dictionary of outdoor species concentrations
             surface_dict = dictionary of surface deposition rates for each species
-            particle_dict = dictionary of particle constants and variables for particle calculations
             num_species = number of species
             
         returns:
@@ -187,7 +185,7 @@ def run_inchem(filename, particles, INCHEM_additional, custom, temp, rel_humidit
         dy_dy=np.zeros((num_species,num_species),dtype=np.float32) #twice as fast as lil_matrix
         
         full_dict={**reaction_rate_dict,**density_dict,**outdoor_dict,**surface_dict,\
-                   **particle_dict,**calc_dict,**timed_dict}
+                   **calc_dict,**timed_dict}
         for k,v in dy_dy_dict.items():
                 dy_dy[v[0],v[1]]=eval(v[2],{},full_dict)
         return dy_dy
@@ -260,8 +258,7 @@ def run_inchem(filename, particles, INCHEM_additional, custom, temp, rel_humidit
         #recalculate particle sums
         if particles == True:
             particle_dict = particle_calcs(part_calc_dict,density_dict)
-        else:
-            particle_dict={}
+            density_dict.update(particle_dict)
             
         #checks time, if between times set for a forced density change the rate
         #is applied to the specific species
@@ -277,12 +274,12 @@ def run_inchem(filename, particles, INCHEM_additional, custom, temp, rel_humidit
         #recalculate reaction rates
         reaction_rate_dict=reaction_eval(reaction_number,J_dict,calc_dict,\
                                          density_dict,dt,reaction_compiled_dict,\
-                                             outdoor_dict,surface_dict,particle_dict,\
+                                             outdoor_dict,surface_dict,\
[…]
