### 2021-05-19 285b0b6
Closes #7
Added a catch to skip the pickle import if there is a version error. The version error will not impact the function of the code but will need to be updated to run the test.
