### 2021-05-19 8f4d682
lowercase module names, #2
Modules with uppercase renamed to lowercase as discussed in issue #2 with `INCHEM_Additional` renamed to `inchem_chemistry` and `Import` renamed to `inchem_import`
