### 2021-05-19 1826b4e
Updated documentation
Updates from multiple issues added to documentation. Closes #4, boolean entries instead of binary integers for choices. Closes #2, lower-case module names.
