+++ manual source files/INCHEM_manual.tex
@@ -32,7 +32,7 @@
 \newcommand{\eg}{e.\,g..}
 \DeclareRobustCommand\cs[1]{\texttt{\char`\\#1}}
 
-\title{Indoor Chemical Model (INCHEM-Py v1.x) \\
+\title{Indoor Chemical Model (INCHEM-Py v1.1) \\
 User Manual\\
 \begin{figure}
     \hspace{-0.25\textwidth}\hspace{0.25in}
@@ -176,9 +176,9 @@ If you're new to Python then the following guide will take you through a quick w
 \linespread{2.0}\selectfont
 \begin{enumerate}
 \setcounter{enumi}{9}
-    \item The full output is saved in "out\_data.pickle". To extract other species concentrations to a csv file for analysis in other software the "INCHEM\_extractor.py" file is used. Double click on this file to open it in the left hand window. 
+    \item The full output is saved in "out\_data.pickle". To extract other species concentrations to a csv file for analysis in other software the "inchem\_extractor.py" file is used. Double click on this file to open it in the left hand window. 
     
-    The following variables can be changed to extract different outputs. Full details of these variables and how to change them can be found in appendix \ref{INCHEM_extractor.py}.
+    The following variables can be changed to extract different outputs. Full details of these variables and how to change them can be found in appendix \ref{inchem_extractor.py}.
                         
             \vspace{1em}
             \begin{minipage}[t]{\linewidth}
@@ -256,14 +256,14 @@ The following files show the folder structure of INCHEM-Py after extraction. If 
  .4 test\_files/.
  .5 custom\_input\_test.txt.
  .5 in\_data.pickle. 
- .5 INCHEM\_test.py. 
+ .5 inchem\_test.py. 
  .5 initial\_test.txt.
- .5 Jacobian\_test.txt.
+ .5 jacobian\_test.txt.
  .5 mcm\_parse\_test.fac.
- .4 Import.py.
- .4 INCHEM\_main.py.
- .4 INCHEM\_test.py.
- .4 INCHEM\_Additional.py.
+ .4 inchem\_import.py.
+ .4 inchem\_main.py.
+ .4 inchem\_test.py.
+ .4 inchem\_chemistry.py.
  .4 initial\_dictionaries.py.
  .4 outdoor\_concentrations.py.
  .4 particle\_input.py.
@@ -271,7 +271,7 @@ The following files show the folder structure of INCHEM-Py after extraction. If 
  .4 reactivity.py.
  .4 surface\_dictionary.py.
  .3 custom\_input.txt.
- .3 INCHEM\_extractor.py.
+ .3 inchem\_extractor.py.
  .3 INCHEMPY\_logo.png.
  .3 INCHEMPY\_logo.svg.
  .3 initial\_concentrations.txt.
@@ -288,7 +288,7 @@ INCHEM-Py consists of a number of input files and a single settings file in whic
 \subsection{Naming conventions}\index{Naming conventions}\label{Naming conventions}
 All species names are in the format given by the MCM download. There is no guide to these but they are self explanatory in most cases. If you are unsure of a species name, then check using \href{http://mcm.leeds.ac.uk/MCM/}{the MCM website} and search for the species using its MCM name or SMILES string.
 
-When using custom inputs it is important that any new species do not share a name with any of the existing species, including those used in the INCHEM\_Additional file inputs.\seealso{\hyperref[custom_input.txt]{custom\_input.txt}}
+When using custom inputs it is important that any new species do not share a name with any of the existing species, including those used in the inchem\_chemistry file inputs.\seealso{\hyperref[custom_input.txt]{custom\_input.txt}}
 
[…]
