### 2021-05-20 fc09b96
Update README.md
`INCHEM_test` to `inchem_test`
