### 2021-05-20 2c8b97a
Idiomatic updates #5
Updates to some sections of code dealing with arrays and lists to more idiomatic python.
