### 2021-05-27 771d864
Timed input spelling check
Added a check that species defined in the timed input actually exist. If they do not then a print statement is issued.
