### 2021-06-03 #3224 @DrDaveShaw
Thanks again all for all of the support and comments, I believe the model is ready for a release alongside the paper as v1.1. @goldmanm how are the additional cases going? Is there anything I can help with?
