### 2021-06-15 #3224 @goldmanm
I made a pull request improving issues with document generation, allowing the test case to run on my computer, and more intuitive messaging when finishing a run. https://github.com/DrDaveShaw/INCHEM-Py/pull/8

I will test the data processing in the next week (feel free to bug me if it isn't done).
