### 2021-06-17 #3224 @DrDaveShaw
I've just merged the pull request, all good changes and very happy with how the code has improved during this process! @goldmanm Thank you very much for taking the time to suggest these changes and for your kind support of the model.

@dhhagan please let me know if there is anything else you need me to do. Obviously the paper refers to version 1.1 which I am ready to release with all of the review changes incorporated. I will do this as soon as I am given the go ahead.

Thank you again for all of your help.
