### 2021-06-18 08cd8b3
Update INCHEM_main.py
Fixed an issue where custom species could be added with the same name as rates or 'RO2'. An additional check has been added to make sure that there are no duplicates to cause confusion or errors.
