### 2021-06-28 #3224 @DrDaveShaw
@dhhagan the final version number is v1.1 and the archive can be found at 10.5281/zenodo.5036374

Thank you very much for all of your help!
