+++ modules/photolysis.py
@@ -590,6 +590,50 @@ def Zixu_photolysis(numba_abs,numba_exp):
                                        "ATTJ76":0.0000,
                                        "ATTJ77":0.0132,
                                        "ATTJ78":0.0132}
+    light_dict["no_sunlight"]= {"ATTJ1":0,
+                                "ATTJ2":0,
+                                "ATTJ3":0,
+                                "ATTJ4":0,
+                                "ATTJ5":0,
+                                "ATTJ6":0,
+                                "ATTJ7":0,
+                                "ATTJ8":0,
+                                "ATTJ11":0,
+                                "ATTJ12":0,
+                                "ATTJ13":0,
+                                "ATTJ14":0,
+                                "ATTJ15":0,
+                                "ATTJ16":0,
+                                "ATTJ17":0,
+                                "ATTJ18":0,
+                                "ATTJ19":0,
+                                "ATTJ20":0,
+                                "ATTJ21":0,
+                                "ATTJ22":0,
+                                "ATTJ23":0,
+                                "ATTJ24":0,
+                                "ATTJ31":0,
+                                "ATTJ32":0,
+                                "ATTJ33":0,
+                                "ATTJ34":0,
+                                "ATTJ35":0,
+                                "ATTJ41":0,
+                                "ATTJ51":0,
+                                "ATTJ52":0,
+                                "ATTJ53":0,
+                                "ATTJ54":0,
+                                "ATTJ55":0,
+                                "ATTJ56":0,
+                                "ATTJ57":0,
+                                "ATTJ70":0,
+                                "ATTJ71":0,
+                                "ATTJ72":0,
+                                "ATTJ73":0,
+                                "ATTJ74":0,
+                                "ATTJ75":0,
+                                "ATTJ76":0,
+                                "ATTJ77":0,
+                                "ATTJ78":0}
     return light_dict
 
 def Zixu_photolysis_compiled():   
