### 2021-07-12 #3224 @kthyng
Congrats on your new publication @DrDaveShaw! And many thanks to editor @dhhagan and reviewers @goldmanm and @khinsen for your time, hard work, and expertise!

(leaving this issue open until the DOI resolves)
